+++ cypress/locators/pod/locators.js
@@ -1,5 +1,6 @@
 // component preview locators
 export const POD_DATA_COMPONENT = '[data-component="pod"]';
+export const POD_BLOCK = "[data-element='block']";
 export const POD_FOOTER = '[data-element="footer"]';
 export const POD_EDIT_ICON = '[data-element="edit"]';
 export const POD_CONTENT = '[data-element="content"]';