+++ src/components/select/simple-select/simple-select.test.js
@@ -4,6 +4,7 @@ import Option from "../option/option.component";
 import OptionRow from "../option-row/option-row.component";
 import OptionGroupHeader from "../option-group-header/option-group-header.component";
 import Icon from "../../icon/icon.component";
+import Box from "../../box";
 import CypressMountWithProviders from "../../../../cypress/support/component-helper/cypress-mount";
 
 import {
@@ -11,16 +12,19 @@ import {
   helpIcon,
   tooltipPreview,
   commonDataElementInputPreview,
+  body,
 } from "../../../../cypress/locators";
 
 import {
   selectText,
   selectInput,
   selectList,
+  selectListWrapper,
   selectOption,
   dropdownButton,
   selectListText,
   multiColumnsSelectListHeader,
+  multiColumnsSelectListHeaderColumn,
   multiColumnsSelectListBody,
   multiColumnsSelectListRow,
   selectListCustomChild,
@@ -319,8 +323,8 @@ const SimpleSelectObjectAsValueComponent = ({ ...props }) => {
 
   return (
     <SimpleSelect
-      id="with-object"
-      name="with-object"
+      id="withObject"
+      name="withObject"
       value={value}
       onChange={onChangeHandler}
       {...props}
@@ -452,6 +456,19 @@ const SimpleSelectEventsComponent = ({ onChange, ...props }) => {
   );
 };
 
+const SimpleSelectWithLongWrappingTextComponent = () => (
+  <Box width={400}>
+    <SimpleSelect name="simple" id="simple" label="label" labelInline>
+      <Option
+        text="Like a lot of intelligent animals, most crows are quite social. 
+        For instance, American crows spend most of the year living in pairs or small family groups.
+        During the winter months, they will congregate with hundreds or even thousands of their peers to sleep together at night."
+        value="1"
+      />
+    </SimpleSelect>
+  </Box>
+);
+
 const testPropValue = "cypress_test";
 
 context("Tests for Simple Select component", () => {
@@ -552,7 +569,7 @@ context("Tests for Simple Select component", () => {
       selectText().click();
       commonDataElementInputPreview().should("have.attr", "readOnly");
       selectText().should("have.attr", "aria-hidden", "true");
-      selectList().should("not.exist");
+      selectList().should("not.be.visible");
     });
 
     it("should render Simple Select as transparent", () => {
@@ -668,7 +685,7 @@ context("Tests for Simple Select component", () => {
       selectList().should("be.visible");
       selectInput().tab();
       selectInput().should("have.attr", "aria-expanded", "false");
-      selectList().should("not.exist");
+      selectList().should("not.be.visible");
     });
 
     it("should close the list with the Esc key", () => {
@@ -678,7 +695,7 @@ context("Tests for Simple Select component", () => {
       selectList().should("be.visible");
       selectText().trigger("keydown", { ...keyCode("Esc") });
       selectInput().should("have.attr", "aria-expanded", "false");
-      selectList().should("not.exist");
+      selectList().should("not.be.visible");
     });
 
     it("should close the list by clicking out of the component", () => {
@@ -686,9 +703,9 @@ context("Tests for Simple Select component", () => {
 
       selectText().click();
       selectList().should("be.visible");
-      cy.get("body").realClick();
+      body().realClick();
       selectInput().should("have.attr", "aria-expanded", "false");
-      selectList().should("not.exist");
+      selectList().should("not.be.visible");
     });
 
     it.each([["downarrow"], ["uparrow"], ["Space"], ["Home"], ["End"]])(
@@ -711,7 +728,7 @@ context("Tests for Simple Select component", () => {
         selectListText(option).click();
         getDataElementByValue("input").should("have.attr", "value", option);
         selectInput().should("have.attr", "aria-expanded", "false");
-        selectList().should("not.exist");
+        selectList().should("not.be.visible");
       }
     );
 
@@ -782,22 +799,40 @@ context("Tests for Simple Select component", () => {
     it("should render list options with multiple columns", () => {
       CypressMountWithProviders(<SimpleSelectMultipleColumnsComponent />);
 
-      const headerColumns = "3";
-      const bodyColumns = "3";
+      const columns = 3;
 
       selectText().click();
       selectList().should("be.visible");
       multiColumnsSelectListHeader()
-        .should("have.length", headerColumns)
+        .should("have.length", columns)
         .and("be.visible");
       multiColumnsSelectListBody()
-        .should("have.length", bodyColumns)
+        .should("have.length", columns)
         .and("be.visible");
       multiColumnsSelectListRow().should(
         "have.css",
         "background-color",
         "rgb(153, 173, 183)"
       );
+    });
+
+    it("should check table header content in list with multiple columns", () => {
+      CypressMountWithProviders(<SimpleSelectMultipleColumnsComponent />);
+
+      const headerCol1 = "Name";
+      const headerCol2 = "Surname";
+      const headerCol3 = "Occupation";
+
+      selectText().click();
+      multiColumnsSelectListHeaderColumn(1)
+        .should("have.text", headerCol1)
+        .and("be.visible");
+      multiColumnsSelectListHeaderColumn(2)
+        .should("have.text", headerCol2)
+        .and("be.visible");
+      multiColumnsSelectListHeaderColumn(3)
+        .should("have.text", headerCol3)
+        .and("be.visible");
     });
 
     it.each([
@@ -888,6 +923,31 @@ context("Tests for Simple Select component", () => {
         selectText().click();
         selectListPosition()
           .should("have.attr", "data-popper-placement", flipPosition)
+          .and("be.visible");
+      }
+    );
+
+    it.each([
+      ["bottom", "0px", "0px", "0px", "0px"],
+      ["top", "600px", "0px", "0px", "0px"],
+      ["right", "200px", "0px", "0px", "900px"],
+      ["left", "600px", "0px", "900px", "0px"],
+    ])(
+      "should render list in %s position with the most space when listPosition is set to auto",
+      (position, top, bottom, left, right) => {
+        CypressMountWithProviders(
+          <SimpleSelectComponent
+            listPlacement="auto"
+            mt={top}
+            mb={bottom}
+            ml={left}
+            mr={right}
+          />
+        );
+
+        selectText().click();
+        selectListPosition()
+          .should("have.attr", "data-popper-placement", position)
           .and("be.visible");
       }
     );
@@ -995,5 +1055,16 @@ context("Tests for Simple Select component", () => {
           });
       }
     );
+  });
+
+  describe("check height of Select list when opened", () => {
+    it("should not cut off any text with long option text", () => {
+      CypressMountWithProviders(<SimpleSelectWithLongWrappingTextComponent />);
+
+      selectText().click();
+      selectListWrapper()
+        .should("have.css", "height", "152px")
+        .and("be.visible");
+    });
   });
 });